Multiple CosmosDb's can be instantiated using clients derived from base type - no more crossover.
