+++ src/Apprentice.Data/CosmosRepositoryBase.cs
@@ -1,0 +1,13 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="CosmosDbRepository.cs" company="Sohara Design Ltd">
+//  This is based on the work of Adam Hockemeyer:
+//   https://github.com/adamhockemeyer/Azure-Functions---CosmosDB-ResourceToken-Broker/blob/master/CosmosDBResourceTokenBroker.Shared/CosmosDBRepository.cs
+// </copyright>
+// <summary>
+//   The cosmos db repository.
+// </summary>
+// --------------------------------------------------------------------------------------------------------------------
+
+namespace ESFA.DAS.ProvideFeedback.Apprentice.Data
+{
+}
