+++ src/Apprentice.Services.FeedbackService/Commands/SaveFeedbackCommand.cs
@@ -6,12 +6,10 @@
     using System.Threading.Tasks;
 
     using ESFA.DAS.ProvideFeedback.Apprentice.Core.Interfaces;
-    using ESFA.DAS.ProvideFeedback.Apprentice.Core.Models;
     using ESFA.DAS.ProvideFeedback.Apprentice.Core.Models.Feedback;
-    using ESFA.DAS.ProvideFeedback.Apprentice.Data.Dto;
     using ESFA.DAS.ProvideFeedback.Apprentice.Data.Repositories;
 
-    using ApprenticeFeedbackDto = Data.Dto.ApprenticeFeedback;
+    using ApprenticeFeedbackDto = ESFA.DAS.ProvideFeedback.Apprentice.Data.Dto.ApprenticeFeedback;
 
     public class SaveFeedbackCommand : ICommandAsync
     {
@@ -21,11 +19,22 @@
 
         private DateTime FinishTime { get; set; }
 
+        private IFeedbackRepository Repository { get; set; }
+
         private List<ApprenticeResponse> Responses { get; set; }
 
         private DateTime StartTime { get; set; }
 
-        private IFeedbackRepository Repository { get; set; }
+        public SaveFeedbackCommand CompletedOn(DateTime time)
+        {
+            this.FinishTime = time;
+            return this;
+        }
+
+        public void Execute()
+        {
+            throw new NotImplementedException();
+        }
 
         public async Task ExecuteAsync(CancellationToken cancellationToken)
         {
@@ -35,27 +44,15 @@
             }
 
             var feedbackDto = new ApprenticeFeedbackDto()
-                {
-                    StartTime = this.StartTime,
-                    FinishTime = this.FinishTime,
-                    Apprentice = this.Apprentice,
-                    Apprenticeship = this.Apprenticeship,
-                    Responses = this.Responses
-                };
+            {
+                StartTime = this.StartTime,
+                FinishTime = this.FinishTime,
+                Apprentice = this.Apprentice,
+                Apprenticeship = this.Apprenticeship,
+                Responses = this.Responses
+            };
 
             await this.Repository.SaveFeedback(feedbackDto);
-        }
-
-        public SaveFeedbackCommand CompletedOn(DateTime time)
-        {
-            this.FinishTime = time;
-            return this;
-        }
-
-        public SaveFeedbackCommand SubmittedBy(Apprentice apprentice)
-        {
-            this.Apprentice = apprentice;
-            return this;
         }
 
         public SaveFeedbackCommand ForApprenticeship(Apprenticeship apprenticeship)
@@ -70,9 +67,9 @@
             return this;
         }
 
-        public SaveFeedbackCommand WithResponses(List<ApprenticeResponse> responses)
+        public SaveFeedbackCommand SubmittedBy(Apprentice apprentice)
         {
-            this.Responses = responses;
+            this.Apprentice = apprentice;
             return this;
         }
 
@@ -82,9 +79,10 @@
             return this;
         }
 
-        public void Execute()
+        public SaveFeedbackCommand WithResponses(List<ApprenticeResponse> responses)
         {
-            throw new NotImplementedException();
+            this.Responses = responses;
+            return this;
         }
     }
 }
